the start of a rewrite of the install scripting framework. i'm aiming for a much simpler single script to handle all the install tasks for the cbench testing tree with the possible exception of binaries (that will depend on the build rework and how it ends up).
the installs for testset files are now symlinks to the
appropriate actual files in tools/templates at the top
of the testing tree. so now there is truly only one
physical start, gen, and output_parse script in the tree.
testset identity is implied by the symlink names via
the find_testset_identity() function or the --testset
parameter.

also, all testsets are now using the new gen_jobs_generic_NEW.pl
script in this install framework.  at some point soon we'll
just promote this gen script to be the only one.

some tickets addressed are #9, #4 (partially)