echo out the jobid in common_header.in
add the --echooutput option to cbench_start_jobs.pl
for the --interactive start mode. this just echos the
job output to the console as well as capturing it to
a .o12345 type file

reworked cbench_functions to support the --echooutput
functionality. the STDIN_FILE stuff should work across
the board now, but i didn't actually test it :).

for the interactive running case i changed both the
stdout and stderr output to be directed to the .o12345
type files instead of .o and .e files