Add some commentary about the custom generation subroutine structure in tools/cbench_gen_jobs.pl.
Add a new way to deal with the walltimes associated with
batch jobs. Cbench has always used a $default_walltime option
defined in cluster.def as the constant walltime assigned
to each job.  Now there is a $walltime_method option in
cluster.def (and as an option to cbench_gen_jobs on the command
line) that selects between two methods: 0 or 1.

# $walltime_method tells cbench_gen_jobs which algorithm to use in
# calculating the walltime for a job:
#    0 - Method 0 is simply a constant value equal to $default_walltime.
#        This has always been the Cbench default behavior.
#    1 - Method 1 is an algorithm that discretely steps up the walltime for
#        each job of increasing runsize (number of processes) by the constant
#        time in minutes controlled by the $walltime_steptime option. In
#        this method, the $default_walltime option is used as the minimum
#        starting walltime from which all others are increased.  The
#        stepped walltimes help schedulers backfill better in general.

The default walltime now has a cbench_gen_jobs command line option as
well.