Since Slurm spools job stdout/stderr output continually into the slurm.oNNNN files, the Cbench output parsing structure appears to have phantom jobs that are in an ERROR state that later disappear. This is because the output parser is looking at output from a live job. Torque/PBS does not behave this way because the .oNNNN file does not show up until the job has completed.
To deal with this intelligently, if cbench_output_parse.pl finds itself
parsing output files from Slurm batch jobs, it will call the slurm_query()
subroutine once to cache the state of Slurm jobs.  Then if the parse
module for a job returns an ERROR status of some sort, the job is cross-
referenced against jobs known to be running in Slurm.  If the job is 
running according to the cached Slurm data, the job is flagged as RUNNING
and not as an error.  For example, here is a snippet from an output parse
run with running jobs:

-------------------------------------------------------------
**DIAG**(lammps/ompi13-intel11/eam.scaled-4ppn-100/slurm.o1012) had ERROR with status STARTED
-------------------------------------------------------------
.**DIAG**(qcd/ompi13-intel11/qcd-4ppn-4/slurm.o1137) is still RUNNING
-------------------------------------------------------------
**DIAG**(cth/ompi13-intel11/amr3doblique-1ppn-1/slurm.o1037) had ERROR with status FATALERROR
-------------------------------------------------------------
**DIAG**(cth/ompi13-intel11/rsrl-1ppn-1/slurm.o1038) had ERROR with status STARTED
**PARSEMATCH**(cth/ompi13-intel11/rsrl-1ppn-1/slurm.o1038) =>  SLURM JOB WALLTIME EXCEEDED
**PARSEMATCH**(cth/ompi13-intel11/rsrl-1ppn-1/slurm.o1038) =>  OMPI says orterun killing job
-------------------------------------------------------------
**DIAG**(cth/ompi13-intel11/amr3doblique-4ppn-8/slurm.o1064) is still RUNNING
-------------------------------------------------------------


slurm_query() was updated a bit to cache Jobid data as well as job name.