Add sPPM to Cbench openapps.
Problem size is statically built into the sppm executable, so any modifications to problem size require a rebuild. Currently, only one binary is built and it has a problem size that uses ~1.2-1.4 GB of memory per thread.

Runtime is set in the inputdeck.cbench file and is currently set to 100 timesteps. Runs of this length and size typically take 1-3+ hours to complete.