adding a testset for the LLNL mpiGraph benchmark.
cbench_run_cmd() in cbench_functions was updated to
support a TEE_FILE variable that allows a job script
to have command line output 'teed' to the file indicated
by TEE_FILE. this was necessary so that the mpigraph_mpigraph.in
job template could tee the output to mpiGraph.out for post
processing by crunch_mpiGraph inside the job itself.

the TEE_FILE thing isn't my favorite thing, but with
all the piping/teeing already going on behind the scenes
in cbench_run_cmd(), it was the only workable way.