RKB:  Add build support for the Scalable HeterOgeneous Computing (SHOC) Benchmark Suite from ORNL (http://ft.ornl.gov/doku/shoc/start).
The Cbench makefile for SHOC relies on the following environment variables to find CUDA and/or OpenCL headers and libraries:

CUDALIB + CUDAINC

and/or

OPENCLLIB + OPENCLINC

I have also taken the liberty of creating a new 'gpu' top-level Cbench make target; so far this group includes SHOC and HPLinpack-gpu.

This commit for SHOC only includes build support, not Cbench job generation and execution.