RKB: Capture the output of `ldd <cbench_binary>` just before executing the job command in the Cbench job script so that the user can recall what version of a binary was used for the test.
The cbench_echo `echo` commands now take "$string" instead of just $string to preserve 
newlines in the `ldd` output. In my quick testing it didn't break other output, but I
could be wrong.

I also took the liberty of combining the if-statements in cbench_check_for_bin() into a 
single if-elif-else statement.