RKB: First pass at using the single node benchmark script to do some NUMA characterization tests. The main addition is run_numa_tests(), a subroutine for running a command using numactl to manage core and memory bindings so that we can test the possible memory-access scenarios of a system.
The two NUMA testing modes determine the numactl options; the 'core-to-node' 
mode runs the program with the process bound to each core and its memory bound to 
each memory node in the system (numactl --physcpubind=x --membind=y); this may be 
thought of as an 'all-to-all' for cores accessing memory in a NUMA system.

The 'node-to-node' mode runs the program with its process bound to any of the cores 
attached to a memory node and with its memory on all possible nodes (numactl 
--cpunodebind=x --membind=y); this may be thought of as 'all-to-all' for memory nodes 
accessing other memory nodes.

Obviously, if core-to-node is used and there are a lot of cores and memory nodes
(e.g. 4P AMD Magny-Cours), then these tests can take quite some time to complete.

The run_numa_tests() subroutine only runs serial benchmarks at this point, but
multi-threaded testing is soon to come.